refactor(l2): remove `ProofCoordinator` concurrent connections limit (#3192)
**Motivation**
In the `ProofCoordinator`'s `handle_listens` function, a semaphore is
used to limit the number of concurrent connections. This constraint can
be removed since it is not necessary and adds complexity to the
codebase.

**Description**
Removes the semaphore and the concurrent connections limit from the
`ProofCoordinator`'s `handle_listens` function.

It also adds the `ConnectionHandler::spawn` method to encapsulate the
`ConnectionHandler` startup logic.