refactor(levm): make substate more accurate and replace ExecutionReport for ContextResult in some places (#3134)
**Motivation**

- `ExecutionResult` isn't accurate for interaction between callframes so
the goal is to replace it for `ContextResult` that has the necessary
data. Also, `Substate` should be as specified in Yellow Paper.

**Description**

- Add logs to substate and remove them from the callframe. They belong
to the substate according to section 6.1 of the [yellow
paper](https://ethereum.github.io/yellowpaper/paper.pdf).
- Replace usage of ExecutionReport in callframes execution for
ContextResult. The former contained data that wasn't necessary and
caused a little bit of confusion. In ContextResult we have only the data
we need: `gas_used`, `output` and `result`.
- Move `is_create` logic to `CallFrame`. So now it is not
`create_op_called`, it is `is_create` and it takes into account external
transactions, not only internal `create`.
- Make functions `handle_opcode_result()` and `handle_opcode_error()`
prettier.
- `finalize_execution()` now returns an `ExecutionReport` given a
`ContextResult`
- Refactor `increase_consumed_gas()`, behavior is still the same but
logic before was kinda repetitive.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3045

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: juanbono <juanbono94@gmail.com>
Co-authored-by: fedacking <francisco.gauna@lambdaclass.com>