chore(l1): change error message shown when loading a pre-merge genesis file (#3111)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
When using a pre-merge genesis.json for importing blocks, which is not
supported by ethrex, the error received was `ParentNotFound`, which
doesn't explain the real problem.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Before merging blocks the genesis.json fork is checked, in case that its
pre Paris return a custom error message.
For doing this new checks were added to the `fork()` function.
<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3102

---------

Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>