chore(levm): improve storage_original_values hashmap and gen_db field names (#3261)
**Motivation**

We can avoid a double lookup with `storage_original_values` by using a
tuple.

I found gen_db field names to be a bit confusing and not giving useful
information, "immutable_cache" is not really a cache but a store for
initial account states, so i changed it. Same with cache.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number