perf(levm): refactor `Stack` to have a fixed size and grow downwards (#3266)
**Motivation**

The stack is currently implemented using a `Vec<U256>` that grows
upwards. Since the stack is limited to a reasonable amount by design
(1024 entries, or 32KiB) it can be converted to a static array along
with an offset and made to grow downwards.

**Description**

This PR:
- Removes stack allocation (and its associated runtime checks) from the
runtime costs.
- Makes the stack grow downwards: better integration with stack
operations.
- Changes the API so that multiple items can be inserted and removed
from the stack at the same time, which is especially useful for opcodes
that handle multiple arguments or return values.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>