perf(levm): refactor levm jump opcodes (#3275)
**Motivation**

The `JUMP` and `JUMPI` opcodes need to check the target address's
validity. This is currently done with a `HashSet` of valid target
addresses, which caused the hashing to become a significant part of the
profiling time when checking for address validity.

**Description**

This PR rewrites the `JUMPDEST` checks so that instead of having a
whitelist, we do the following:
- Check the program bytecode directly. The jump target's value should be
a `JUMPDEST`.
- Check a blacklist of values 0x5B (`JUMPDEST`) that are NOT opcodes
(they are part of push literals).

The blacklist is not a `HashMap`, but rather a sorted slice that can be
checked efficiently using the binary search algorithm, which should
terminate on average after the first or second step.

Rational: After extracting stats of the first 10k hoodi blocks, I found
out that...
- There are almost 60 times more `JUMPDEST` than values 0x5B in push
literals.
- On average, there are less than 2 values in the blacklist. If we were
to use a whitelist as before, there would be about 70 entries on
average.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Related to #3305