refactor(l1, l2): remove `l2` feature flag from `rpc` crate (first proposal) [RFC] (#3330)
**Motivation**

The main goal is to decouple L2 logic from the crate `rpc`. 

This PR proposes an initial solution to this.

**Description**

This solution extends `ethrex-rpc` by wrapping it where needed, and
adding new data structures. I think a better solution can be achieved
from this one.

- Adds `ethrex-l2-rpc` crate in `crates/l2/networking/rpc`.
- Moves L2 logic from `ethrex-rpc` to `ethrex-l2-rpc`.
- Refactors some functions in `ethrex-rpx` to be reusable in
`ethrex-l2-rpc`.
- Exposes some functions and types from `ethrex-rpc`.
- Updates `cmd/ethrex` with this changes and moves L2 initializers from
`cmd/ethrex/initializers.rs` to `cmd/ethrex/l2/initializers.rs`.

**Pros and Cons**

| Pros | Cons|
| --- | --- |
| L2 logic is decoupled from `ethrex-rpc` | L2 logic is decoupled from
`ethrex-rpc` by wrapping `ethrex-rpc` functions and duplicating some
types to extend them |
| `ethrex-rpc` logic is reused by `ethrex-l2-rpc` | Some types and
functions were exposed in `ethrex-rpc` public API |

Despite the cons, this could be an acceptable first iteration to a
proper solution as this implementation highlights somehow which parts of
the rpc crate need to be abstracted.

**Next Steps**

- Remove `l2` feature flag from `cmd/ethrex`.
- Move `crates/networking/rpc/clients` module.
- Cleanup `ethrex-rpc` public API.
- The next iteration should include an more general abstraction of the
RPC API implementation so it can be extended organically instead of by
wrapping it.