feat(l1): embed genesis of public networks in binary (#3372)
**Motivation**

Currently, users need to download the genesis to be able to run a
network without being in the repo's root.

**Description**

This PR embeds the genesis file of known public networks inside the
compiled binary. It has the downside of increasing binary size from 23.6
Mb to 24.7 Mb.

Further code simplifications are left for other PRs.

In the future, a possible improvement could be to parse each genesis
file before embedding it in the binary. Now we just embed the plain
JSON.

Related #3292