perf(levm): refactor `CacheDB` to use more efficient APIs (#3259)
**Motivation**

The cache db is a bunch of functions that accept a state object as an
argument. This is confusing since those are not methods, but functions,
which also do stuff that the state object already supports natively (not
to mention the duplicated function).

**Description**

Remove the `cache.rs` file and use the state object directly. Move stuff
to more relevant places to fix borrow issues.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number