refactor(l1, l2): decouple L2 metrics logic from L1's & remove l2 feature flag from `ethrex-blockchain` (#3371)
> [!WARNING]
> Merge after https://github.com/lambdaclass/ethrex/pull/3367

**Motivation**

To completely remove the `l2` feature flag from `cmd/ethrex` in favor of
having a single binary for running ethrex (L1 and L2), there are some
local dependencies from which to remove this feature first. These are:

1. `ethrex-vm`.
2. `ethrex-levm`.
3. `ethrex-blockchain`. 

1 and 2 are removed in https://github.com/lambdaclass/ethrex/pull/3367,
and 3 is meant to be removed in this PR.

**Description**

Decouples the L2 metrics logic from the L1's, allowing to remove the use
of the `l2` feature flag from the crate `ethrex-blockchain`.

- Creates a `crates/blockchain/metrics/l2` module with `metrics.rs` and
`api.rs` submodules.
- Makes use of this new module in `cmd/ethrex`.
- Removes `l2` feature flag from `ethrex-blockchain` crate.
- Removes the import of `ethrex-blockchain/l2` where needed.