chore(l2): re-enable risc0 (#3172)
**Motivation**

risc0 support was temporarily deprecated because of incompatible
versions between some required ethrex's dependencies and the risc0
toolchain.

Now the toolchain uses a newer version so this problem should be solved,
but the backend needs some maintenance to get it working again.

**Description**

- update risc0 to latest version
- update risc0's build script for the new version
- refactor kzg verification into ethrex-common
- support kzg verification with both kzg-rs and c-kzg (sp1 is only
compatible with kzg-rs, risc0 only with c-kzg)
- fix wrong public inputs encoding
- fix wrong image id encoding
- add risc0 verification key (also called image id) as a contract
variable
- add risc0 lint job and refactor jobs for other backends
- add docs for local testing (deployment of risc0 contracts)

Closes #2145