chore(l1,levm): improve eest consume tests report and add Makefile (#3310)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

- Created Makefile commands for running eest tests. Note that by default
we ran tests for forks after Paris (including this one).
- Changed the hive report generator. Now (if it works, I haven't tested
it yet) we create results that are actually accurate. Before we were
saying that Cancun results were from tests/cancun and that was not
necessarily true, in that folder there can be tests that are from other
forks too. What determines the fork that we use to run the test is the
`fork_(fork name)` in the test name. That is the explanation for the
regex and related changes.

[Example
run](https://github.com/lambdaclass/ethrex/actions/runs/15928920747)

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3395