refactor(core): remove l2 feature from `cmd/ethrex` achieving a single binary for both L1 and L2 (#3381)
> [!WARNING]
> Merge after #3371

**Motivation**

We want `ethrex` to be a single binary for both running L1 and L2. For
this, we need to remove the `l2` feature flag from `cmd/ethrex`, a work
that includes removing the same feature flag from the following crates:

1. `ethrex-vm`.
2. `ethrex-levm`.
3. `ethrex-blockchain`. 

1 and 2 are removed in https://github.com/lambdaclass/ethrex/pull/3367,
and 3 is removed in #3371.

**Description**

Removes `l2` feature flag from `cmd/ethrex`