feat(l2): proxied l2 system contracts (#3421)
**Motivation**

We want to be able to upgrade L2 system contracts.

**Description**

This makes it so that the L2 contracts themselves are proxies. Their
initial implementations are kept in the genesis for ease of deployment
and to avoid keeping them empty in the first blocks.

Since the proxies need to be embedded in the genesis, they can't be
deployed with a constructor, so their
[ERC-1967](https://eips.ethereum.org/EIPS/eip-1967) slots are set
directly.

A function is added to the L1 CommonBridge to allow upgrading the L2
contracts. A special address (0xf000) is used to authenticate the
upgrade.

Closes #3345